requirements.txt
flask==2.2.3
python-dotenv==1.0.0
flask-pymongo==2.3.0
flask-wtf==1.1.1
flask-session==0.4.0
